move ct keyring responsibilities to trustedroot
