+++ sigstore/_internal/sct.py
@@ -36,8 +36,8 @@ from cryptography.x509.certificate_transparency import (
 )
 from cryptography.x509.oid import ExtendedKeyUsageOID
 
-from sigstore._internal.ctfe import CTKeyring
 from sigstore._internal.trustroot import (
+    CTKeyring,
     KeyringError,
     KeyringLookupError,
     KeyringSignatureError,
